feat(compiler): Cleaning up auto generating of unique ids for types
