+++ module/manifest/__index__.ts
@@ -7,6 +7,5 @@ export * from './src/manifest-index';
 export * from './src/runtime';
 export * from './src/package';
 export * from './src/util';
-export * from './src/trie';
 export * from './src/file';
 export * from './src/types';
