+++ module/manifest/src/util.ts
@@ -122,4 +122,41 @@ export class ManifestUtil {
       }
     };
   }
+
+  /**
+   * Efficient lookup for path-based graphs
+   */
+  static lookupTrie<T>(
+    inputs: T[], getPath: (v: T) => string[], validateUnknown?: (pth: string[]) => boolean
+  ): (path: string[]) => T | undefined {
+    type TrieNode<T> = { value?: T, subs: Record<string, TrieNode<T>> };
+    const root: TrieNode<T> = { subs: {} };
+    for (const item of inputs) {
+      const pth = getPath(item);
+      let node = root;
+      for (const sub of pth) {
+        node = node.subs[sub] ??= { subs: {} };
+      }
+      node.value = item;
+    }
+
+    return pth => {
+      let node = root;
+      let value = node.value;
+      let i = 0;
+
+      for (const sub of pth) {
+        i += 1;
+        if (node) {
+          node = node.subs[sub];
+          value = node?.value ?? value;
+        } else if (validateUnknown && !node && !validateUnknown(pth.slice(0, i))) {
+          value = undefined;
+          break;
+        }
+      }
+
+      return value;
+    }
+  }
 }
